pyproject.toml
[tool.poetry]
name = "domydrafter"
version = "0.1.0"
description = ""
authors = ["Etienne Soucy <soucy.etienne@courrier.uqam.ca>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.8.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
